requirements.txt
# ═══════════════════════════════════════════════════════════
# Saudi AI-ERP SaaS Platform — Python Dependencies
# ═══════════════════════════════════════════════════════════

# ─── Core Framework ───────────────────────────────────────
Django>=4.2,<5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
django-environ>=0.11
django-extensions>=3.2
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2
django-allauth[socialaccount]>=0.61
PyJWT[crypto]>=2.8

# ─── Multi-Tenancy ───────────────────────────────────────
django-tenants>=3.6

# ─── Database ─────────────────────────────────────────────
psycopg2-binary>=2.9
dj-database-url>=2.1

# ─── ASGI / WebSocket / Channels ─────────────────────────
daphne>=4.1
channels>=4.0
channels-redis>=4.2

# ─── Cache ────────────────────────────────────────────────
django-redis>=5.4

# ─── Async Tasks ──────────────────────────────────────────
celery>=5.3
redis>=5.0
django-celery-beat>=2.6
django-celery-results>=2.5

# ─── State Machine ───────────────────────────────────────
django-fsm>=3.0

# ─── ZATCA / E-Invoicing ─────────────────────────────────
requests>=2.31
zeep>=4.2.1
xmlsec>=1.3
lxml>=5.1
cryptography>=42.0
qrcode>=7.4
hijri-converter>=2.3

# ─── AI / LLM ────────────────────────────────────────────
anthropic>=0.39
# ollama  # Uncomment for air-gapped deployments

# ─── Reports & Exports ───────────────────────────────────
reportlab>=4.1
openpyxl>=3.1

# ─── Media & Storage ─────────────────────────────────────
Pillow>=10.2
django-storages[boto3]>=1.14
boto3>=1.34

# ─── Production Server ───────────────────────────────────
gunicorn>=21.2
whitenoise>=6.6

# ─── Import/Export ────────────────────────────────────────
django-import-export>=3.3

# ─── Observability ────────────────────────────────────────
sentry-sdk[django,celery]>=2.0
